feat: Migrate to module

+++ scripts/backup.sh
@@ -5,14 +5,9 @@
 ###
 set -euxo pipefail
 
-###
-## Run a command via mcrcon
-##
-## This will run the given command over mcrcon, automatically substituting in the password.
-###
-run_cmd() {
-    mcrcon -p $(cat @rconPasswordFile@) "$@"
-}
+# Load in rcon variables
+export MCRCON_PORT="@rconPort@"
+export MCRCON_PASS="$(cat @rconPasswordFile@)"
 
 ###
 ## Announce something
@@ -24,10 +19,21 @@ run_cmd() {
 ## 1. The text to format and announce
 ###
 announce() {
-    run_cmd "tellraw @a [\"\",{\"text\":\"[\",\"bold\":true,\"color\":\"red\"}\
+    mcrcon "tellraw @a [\"\",{\"text\":\"[\",\"bold\":true,\"color\":\"red\"}\
         ,{\"text\":\"ANN\",\"bold\":true,\"color\":\"dark_red\"},{\"text\":\"]\",\"bold\":true,\
         \"color\":\"red\"},{\"text\":\" - \",\"bold\":true,\"color\":\"dark_gray\"},{\"text\":\
         \"$1\",\"underlined\":true,\"color\":\"gold\"}]"
+}
+
+###
+## Create a borg archive with the given name and of the given directory
+##
+## # Arguments
+## 1. The name of the archive
+## 2. The directory to archive
+###
+backup() {
+    borg create --compression zstd,16 --list -s @backupDirectory@/minecraft::$1 $2
 }
 
 ###
@@ -57,20 +63,22 @@ announce "Server backup starting, there may be some lag."
 # Reset the timer
 SECONDS=0
 # First save everything, then turn off chunk saving
-run_cmd "save-off"
-run_cmd "save-all flush"
+mcrcon "save-off"
+mcrcon "save-all flush"
 
 # Sleep 3 seconds and run a sync just to make sure everything gets written, we don't want any of
 # those pesky "modified while backing up" errors
 sleep 3
 sync
 
-# Do the borg backup with the following options
-borg create --compression zstd,16 --list -s @backupDirectory@/minecraft::$(date "+%Y%m%d-%H%M%S")\
-    /var/minecraft/server
+timestamp=$(date "+%Y%m%d-%H%M%S")
+# Backup the world
+backup "world-$timestamp" /var/minecraft/world
+# Backup the server files
+backup "server-$timestamp" /var/minecraft/server
 
 # After we are done with the backup, turn saving back on
-run_cmd "save-on"
+mcrcon "save-on"
 # Tell the players that we are done
 secs=$SECONDS
 # Formatting
@@ -82,28 +90,30 @@ fi
 announce "Uploading backup.";
 
 ###
-## Copy our backup to backblaze
+## Copy our backup to backblaze, if enabled
 ###
-# Reset the seconds timer
-SECONDS=0
-# Update the rlone configuraion
-mkdir -p ~/.config/rclone
-[ -f ~/.config/rclone/rclone.conf  ] && rm ~/.config/rclone/rclone.conf
-cat <<EOF > ~/.config/rclone/rclone.conf
+
+if @b2Enable@; then
+    # Reset the seconds timer
+    SECONDS=0
+    # Update the rlone configuraion
+    mkdir -p ~/.config/rclone
+    [ -f ~/.config/rclone/rclone.conf  ] && rm ~/.config/rclone/rclone.conf
+    cat <<EOF > ~/.config/rclone/rclone.conf
 [b2]
     type = b2
     account = @b2AccountID@
     key = $(cat @b2KeyFile@)
 EOF
-# Make the backup
-rclone copy @backupDirectory@ b2:@b2Bucket@/ -P --transfers 16
-# Report on the upload status
-secs=$SECONDS
-# Formatting
-if [ $secs -ge 60 ]; then
-    announce "Server backup uploaded in $(printf '%dm:%ds' $((secs%3600/60)) $((secs%60)))."
-else
-    announce "Server backup uploaded in $(printf '%ds' $((secs%60)))."
+    # Make the backup
+    rclone copy @backupDirectory@ b2:@b2Bucket@/ -P --transfers 16
+    # Report on the upload status
+    secs=$SECONDS
+    # Formatting
+    if [ $secs -ge 60 ]; then
+        announce "Server backup uploaded in $(printf '%dm:%ds' $((secs%3600/60)) $((secs%60)))."
+    else
+        announce "Server backup uploaded in $(printf '%ds' $((secs%60)))."
+    fi
+
 fi
-# FIN
-announce ""
